pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "jira-agent"
version = "0.1.3"
description = "Jira automation agent with Streamlit dashboard and API"
readme = "README.md"
requires-python = ">=3.8"
license = { text = "MIT" }
authors = [{ name = "Saad Rehman", email = "saadrehman@outlook.com" }]
dependencies = [
  "jira>=3.5.2",
  "openai>=1.12.0",
  "python-dotenv>=1.0.0",
  "requests>=2.31.0",
  "pydantic>=2.5.0",
  "streamlit>=1.28.0",
  "plotly>=5.17.0",
  "pandas>=2.1.0",
  "phidata>=0.5.0",
]


[project.scripts]
jir-agent-dashboard = "run_dashboard:main"
jir-agent-api = "uvicorn:main"  # run as: uvicorn main:app --reload

[tool.setuptools]
packages = { find = { where = ["."], include = ["*"], exclude = ["venv*", "tests*", "*.zip", "teams_app*"] } }
include-package-data = true
py-modules = [
  "analytics",
  "analytics_tools",
  "config",
  "dashboard_app",
  "enhanced_streamlit_app",
  "install",
  "jira_client",
  "llm_client",
  "main",
  "phi_jira_agent_final",
  "run_dashboard",
  "simple_jira_tools",
]


